rag_consistency_study/requirements.txt
langchain-core==1.3.3
langchain-openai==1.2.1
langchain-community==0.4.1
faiss-cpu==1.13.2
pandas==3.0.1
openpyxl==3.1.5
python-dotenv==1.2.2
requests==2.32.5
pydantic==2.12.5
